Convert p4 labels into git tags
This first requires working out which changelist number was tagged.

Then we work out the git branch, to disambiguate which commit to tag.
This was necessary for Open Watcom, because the 2.2.0 and 2.2.1 tags
both tagged different views of the same revision.

We also provide a manual label for Open Watcom v1.0.0. This was
missing from the Perforce depot.